+++ docs/_pages/releases.md
@@ -13,6 +13,7 @@ sidebar:
 * Added official support for .NET Core 3.0
 * Added `WithOffset` extension method on `DateTime` for easier creation of `DateTimeOffset` objects.
 * Added `collectionOfStrings.Should().NotContainMatch()` to assert that the collection does not contain a string that matches a wildcard pattern 
+* The `Using`/`When` option on `BeEquivalentTo` will now use the conversion rules when trying to match the predicate.
 
 **Fixes**
 * Reported actual value when it contained `{{{{` or `}}}}` (#1223)